Open up the Z3 api.
