+++ backends/p4tools/common/core/z3_solver.h
@@ -44,6 +44,14 @@ class Z3Solver : public AbstractSolver {
 
     std::optional<bool> checkSat(const std::vector<const Constraint *> &asserts) override;
 
+    /// Z3Solver specific checkSat function. Calls check on the input z3::expr_vector.
+    /// Only relies on the incrementality mode of the Z3 solver.
+    std::optional<bool> checkSat(const z3::expr_vector &asserts);
+
+    /// Z3Solver specific checkSat function. Calls check on the solver.
+    /// Only useful in incremental mode.
+    std::optional<bool> checkSat();
+
     [[nodiscard]] const SymbolicMapping &getSymbolicMapping() const override;
 
     void toJSON(JSONGenerator & /*json*/) const override;
@@ -73,10 +81,13 @@ class Z3Solver : public AbstractSolver {
     /// In incremental state, all active assertions are reapplied after resetting.
     void clearMemory();
 
- private:
+    /// Adds a Z3 assertion to the solver context.
+    void asrt(const z3::expr &assert);
+
     /// Inserts an assertion into the topmost solver context.
     void asrt(const Constraint *assertion);
 
+ private:
     /// Converts a P4 type to a Z3 sort.
     z3::sort toSort(const IR::Type *type);
 
@@ -101,6 +112,9 @@ class Z3Solver : public AbstractSolver {
     //             that is greater than or equal to @asrtIndex.
     /// Helps to restore a state of incremental solver in a constructor.
     void addZ3Pushes(size_t &chkIndex, size_t asrtIndex);
+
+    /// Helper function which converts a z3::check_result to a std::optional<bool>.
+    static std::optional<bool> interpretSolverResult(z3::check_result result);
 
     /// The underlying Z3 instance.
     z3::solver z3solver;
@@ -130,6 +144,104 @@ class Z3Solver : public AbstractSolver {
     std::optional<unsigned> timeout_;
 
     DECLARE_TYPEINFO(Z3Solver, AbstractSolver);
+};
+
+/// Translates P4 expressions into Z3. Any variables encountered are declared to a Z3 instance.
+class Z3Translator : public virtual Inspector {
+ public:
+    /// Creates a Z3 translator. Any variables encountered during translation will be declared to
+    /// the Z3 instance encapsulated within the given solver.
+    explicit Z3Translator(Z3Solver &solver);
+
+    /// Handles unexpected nodes.
+    bool preorder(const IR::Node *node) override;
+
+    /// Translates casts.
+    bool preorder(const IR::Cast *cast) override;
+
+    /// Translates constants.
+    bool preorder(const IR::Constant *constant) override;
+    bool preorder(const IR::BoolLiteral *boolLiteral) override;
+    bool preorder(const IR::StringLiteral *stringLiteral) override;
+
+    /// Translates variables.
+    bool preorder(const IR::SymbolicVariable *var) override;
+
+    // Translations for unary operations.
+    bool preorder(const IR::Neg *op) override;
+    bool preorder(const IR::Cmpl *op) override;
+    bool preorder(const IR::LNot *op) override;
+
+    // Translations for binary operations.
+    bool preorder(const IR::Equ *op) override;
+    bool preorder(const IR::Neq *op) override;
+    bool preorder(const IR::Lss *op) override;
+    bool preorder(const IR::Leq *op) override;
+    bool preorder(const IR::Grt *op) override;
+    bool preorder(const IR::Geq *op) override;
+    bool preorder(const IR::Mod *op) override;
+    bool preorder(const IR::Add *op) override;
+    bool preorder(const IR::Sub *op) override;
+    bool preorder(const IR::Mul *op) override;
+    bool preorder(const IR::Div *op) override;
+    bool preorder(const IR::Shl *op) override;
+    bool preorder(const IR::Shr *op) override;
+    bool preorder(const IR::BAnd *op) override;
+    bool preorder(const IR::BOr *op) override;
+    bool preorder(const IR::BXor *op) override;
+    bool preorder(const IR::LAnd *op) override;
+    bool preorder(const IR::LOr *op) override;
+    bool preorder(const IR::Concat *op) override;
+
+    // Translations for ternary operations.
+    bool preorder(const IR::Mux *op) override;
+    bool preorder(const IR::Slice *op) override;
+
+    /// @returns the result of the translation.
+    z3::expr getResult();
+
+    /// Translate an P4C IR expression and return the Z3 equivalent.
+    z3::expr translate(const IR::Expression *expression);
+
+ private:
+    /// Function type for a unary operator.
+    using Z3UnaryOp = z3::expr (*)(const z3::expr &);
+
+    /// Function type for a binary operator.
+    using Z3BinaryOp = z3::expr (*)(const z3::expr &, const z3::expr &);
+
+    /// Function type for a ternary operator.
+    using Z3TernaryOp = z3::expr (*)(const z3::expr &, const z3::expr &, const z3::expr &);
+
+    /// Handles recursion into unary operations.
+    ///
+    /// @returns false.
+    bool recurseUnary(const IR::Operation_Unary *unary, Z3UnaryOp f);
+
+    /// Handles recursion into binary operations.
+    ///
+    /// @returns false.
+    bool recurseBinary(const IR::Operation_Binary *binary, Z3BinaryOp f);
+
+    /// Handles recursion into ternary operations.
+    ///
+    /// @returns false.
+    bool recurseTernary(const IR::Operation_Ternary *ternary, Z3TernaryOp f);
+
+    /// Rewrites a shift operation so that the type of the shift amount matches that of the number
+    /// being shifted.
+    ///
+    /// P4 allows shift operands to have different types: when the number being shifted is a bit
+    /// vector, the shift amount can be an infinite-precision integer. This rewrites such
+    /// expressions so that the shift amount is a bit vector.
+    template <class ShiftType>
+    const ShiftType *rewriteShift(const ShiftType *shift) const;
+
+    /// The output of the translation.
+    z3::expr result;
+
+    /// The Z3 solver instance, to which variables will be declared as they are encountered.
+    std::reference_wrapper<Z3Solver> solver;
 };
 
 }  // namespace P4Tools
